rebasing master branch fixing conflicts
