+++ tests/getServerSideProps.test.ts
@@ -6,12 +6,10 @@ import { enableFetchMocks } from "jest-fetch-mock";
 import jwt from "jsonwebtoken";
 import { Socket } from "net";
 
+import { ReceivedIdentityTypes } from "../src/@types/Identity";
+import { ProviderUser } from "../src/@types/ProviderUser";
 import { config } from "../src/config";
 import { getServerSideProps } from "../src/pages/account";
-import {
-  mockedSortedResponse,
-  mockedResponse,
-} from "./__mocks__/managementResponse";
 
 enableFetchMocks();
 
@@ -50,6 +48,34 @@ describe("getServerSideProps", () => {
     mockedContext.req.rawHeaders = [];
   });
 
+  const mockedResponse: { data: { provider: ProviderUser } } = {
+    data: {
+      provider: {
+        id: "4a5f8589-0d91-4a69-924a-6f227a69666d",
+        name: "Mocked Provider",
+        user: {
+          id: "299d268e-3e19-4486-9be7-29c539d241ac",
+          identities: [
+            {
+              id: "7f8d168a-3f65-4636-9acb-7720a212680e",
+              primary: true,
+              status: "validated",
+              type: ReceivedIdentityTypes.PHONE,
+              value: "0123456789",
+            },
+            {
+              id: "8f79dcc1-530b-4982-878d-33f0def6a7cf",
+              primary: true,
+              status: "validated",
+              type: ReceivedIdentityTypes.EMAIL,
+              value: "test@test.test",
+            },
+          ],
+        },
+      },
+    },
+  };
+
   it("should redirect to the login flow if there are no session", async () => {
     fetch.once(JSON.stringify(mockedResponse));
 
@@ -69,6 +95,27 @@ describe("getServerSideProps", () => {
   });
 
   it("should get the user-id from the session and call management GraphQL endpoint", async () => {
+    const mockedSortedResponse = {
+      emailIdentities: [
+        {
+          id: "8f79dcc1-530b-4982-878d-33f0def6a7cf",
+          primary: true,
+          status: "validated",
+          type: ReceivedIdentityTypes.EMAIL,
+          value: "test@test.test",
+        },
+      ],
+      phoneIdentities: [
+        {
+          id: "7f8d168a-3f65-4636-9acb-7720a212680e",
+          primary: true,
+          status: "validated",
+          type: ReceivedIdentityTypes.PHONE,
+          value: "0123456789",
+        },
+      ],
+    };
+
     const JWT = jwt.sign(
       mockedJWTPayload,
       config.connectApplicationClientSecret,
